refactor: restructure todo components and optimize modal rendering
- Separated TodoActionButtons and DeleteTodoModal into independent components.
- Organized folder structure:
    - src/components/ui (shared elements)
    - src/components/todo (logic-specific components)
    - src/components/modals (project-wide modals)
- Implemented conditional rendering (short-circuit) for modals to prevent unnecessary DOM mounting.
- Improved component maintainability and initial render performance.

- Розділено TodoActionButtons та DeleteTodoModal на незалежні компоненти.
- Організована структура папок:
- src/components/ui (спільні елементи)
- src/components/todo (логічно-специфічні компоненти)
- src/components/modals (модальні вікна всього проєкту)
- Реалізовано умовний рендеринг (коротке замикання) для модальних вікон, щоб запобігти непотрібному монтуванню DOM.
- Покращено підтримку компонентів та початкову продуктивність рендерингу.
