+++ src/components/modals/ModalActions.jsx
@@ -1,4 +1,4 @@
-import Button from "./Button";
+import Button from "../ui/Button";
 
 const ModalActions = ({
   onCancel,
